Extracts icons functions into Icons module
